generalize isl_multi_val_read_from_str to cover isl_multi_id_read_from_str
