+++ Makefile.am
@@ -456,6 +456,7 @@ EXTRA_DIST = \
 	isl_multi_nan_templ.c \
 	isl_multi_param_templ.c \
 	isl_multi_product_templ.c \
+	isl_multi_read_no_explicit_domain_templ.c \
 	isl_multi_splice_templ.c \
 	isl_multi_tuple_id_templ.c \
 	isl_multi_zero_templ.c \
